combined spotifyAuth.jsx and loginBtn.jsx to streamline process. Also added some basic styling
