+++ client/src/assets/styles.css
@@ -1,3 +1,21 @@
+body {
+  background: linear-gradient(to right, #100d29, #231f4b, #100d29);
+  color: var(--text-color);
+    margin: 0;
+    padding: 0;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+      sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+  }
+  
+ 
+  main {
+    color: #ec0aa1; 
+    padding: 1rem;
+  }
+
 .navbar {
     display: flex;
     justify-content: space-between;
@@ -21,3 +39,34 @@
     padding: 1rem 0;
 }
 
+/* Neon text effect */
+.neon-text {
+    color: #3f1919;
+    text-shadow:
+      0 0 5px #c8c8c8,
+      0 0 10px #c8c8c8,
+      0 0 20px #c8c8c8,
+      0 0 40px #ff00ff,
+      0 0 80px #ff00ff,
+      0 0 90px #ff00ff,
+      0 0 100px #ff00ff,
+      0 0 150px #ff00ff;
+  }
+
+
+/* Neon button effect */
+.neon-button {
+    border: 1px solid #05f5f5;
+    background-color: #070816;
+    color: #55bb47;
+    text-shadow: 0 0 5px rgba(196, 196, 250, 0.75);
+    transition: all 0.3s ease;
+  }
+
+  .neon-button:hover {
+    box-shadow:
+      0 0 5px #c8c8c8,
+      0 0 10px #c8c8c8,
+      0 0 15px #ff00ff,
+      0 0 20px #ff00ff;
+  }
