+++ client/src/components/layout.jsx
@@ -13,3 +13,4 @@ const Layout = ({ children }) => {
 };
 
 export default Layout;
+
